+++ app/models/folio/console_ability.rb
@@ -9,7 +9,12 @@ class Folio::ConsoleAbility
     case account.role
     when "manager"
       can :manage, :all
+      cannot :manage, :sidekiq
       cannot :manage, Folio::Account, role: "superuser"
+      cannot :manage, Folio::Account, role: "administrator"
+    when "administrator"
+      can :manage, :all
+      cannot :manage, :sidekiq
     when "superuser"
       can :manage, :all
     end