+++ config/routes.rb
@@ -159,4 +159,11 @@ Folio::Engine.routes.draw do
                                   constraints: { name: /.*/ }
 
   get "/sitemaps/:id.:format(.:compression)", to: "sitemaps#show"
+
+  require "sidekiq/web"
+  require "sidekiq/cron/web"
+
+  authenticate :account, lambda { |account| account.can_manage_sidekiq? } do
+    mount Sidekiq::Web => "/sidekiq"
+  end
 end